+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/media-selection/main.js
@@ -85,14 +85,6 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
         },
 
         /**
-         * raised when selection is changed
-         * @event sulu.media-selection.document-selection.data-changed
-         */
-        DATA_CHANGED = function() {
-            return createEventName.call(this, 'data-changed');
-        },
-
-        /**
          * returns normalized event names
          */
         createEventName = function(postFix) {
@@ -243,8 +235,6 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
             }
 
             this.$el.find('#selected-images-count').html(template);
-            this.sandbox.emit(DATA_CHANGED.call(this));
-
         },
 
         /**
@@ -274,7 +264,6 @@ define(['sulumedia/collection/collections', 'sulumedia/model/collection'], funct
                     }.bind(this)
                 });
             }.bind(this));
-
 
             // save the collection where new media should be uploaded
             this.sandbox.on(